Add additional API tests to confirm functionality
The changes here add additional tests to the system to ensure all of
the required APIs are up, available, and are returning data as expected.
To ensure the tests run the `q_mem` override value has been set a little
higher, which will allow logstash to start (before it would result in a
heap size of 115M which is not enough to start logstash).

Change-Id: Ie640f609ef345e9b4908eb17a03be5390b72b186
Signed-off-by: Kevin Carter <kevin.carter@rackspace.com>