multi-noda-aio loadbalancer routing issues fix
The loadbalancer vm is configured to utilize br-dhcp for the external
vip and br-mgmt for the internal vip. Users end up accessing these
from various networks depending on the dev/testing going on. For
example if you utilize the flat network for vms that need to hit the
external api, the calls fail. This is due to traffic coming in on
br-dhcp then going back out on br-flat. This change is to remove the
unused interfaces to cut down on routing issues in the load balancer.

Change-Id: Ic3c3bd18e63943ff288692da20d449852a316b5e
Closes-Bug: #1795710