MNAIO: Log aria2c results
In order to aid in the troubleshooting of download failures,
we set aria2c to produce a log in a configurable location.
Rather than use the default log level of 'debug', which logs
so much information that it affects the host performance due
to the massive amount of IOPS, we set it to 'notice' which
should be sufficient information for debugging most situations.

Change-Id: I67a41026f8e62562a9f3d9c08512f5da2a27f3dc