MNAIO: Do not use apt-cacher-ng on the MNAIO host
If all VM's are down, then the IP Address that apt-cacher-ng is
listening on will also be down, resulting in the host not being
capable of reaching it for any apt activities.

Rather than try to work around this, we simply let the host access
its upstream repositories directly given that the cache is primarily
for the purpose of speeding up the VM deployment.

Change-Id: I12f474321e9a9302d361907a882051c963ccf045