MNAIO: Use the OSA inventory and plugins where possible
There are times when it's useful to run playbooks against the
OSA inventory, but having to do so via infra1 is a bit of a
pain. Rather than do that we add a script which fetches the
OSA inventory file from infra1, and we also bootstrap the OSA
plugins repo into the default ansible plugins path so that it
is possible to run playbooks against the containers in the VM's
from the host without needing to do anything else.

Change-Id: I751af19841ea825d21dd67f4343674b8522b689f