MNAIO: Copy SSH keys into containers if they are present
When starting an imaged environment, the SSH keys on the host do
not match the SSH keys in the containers. In this patch we ensure
that if the 'all_containers' group is present, then the deploy-vms
playbook will wait for them to come up and copy the new keys into
them. It also adds the new public key into the authorized_keys.
This ensures that the whole environment is ready to be used just
as it was before imaging.

We also remove the 'port' argument given to the wait_for_connection
module because the argument is invalid.

Change-Id: Iff0a3327c2031e0dd977e8e403b417e495fee14f