MNAIO: Remove legacy os-infra_hosts group
The legacy group 'os-infra_hosts' is not actually part
of the infrastructure - it includes all the openstack
infrastructure groups (keystone, nova, neutron, glance,
heat, swift). This group's use is unnecessary because
all the other groups included are represented with
individual options.

Having this in the osa_enable_infra conditional also
means that if you mean to disable swift/heat, it does
not work and you end up with a broken deployment.

Change-Id: Icd80fd96aad713372b1fe21752799d56ada3dac4