MNAIO: Implement branch var for pinning Ansible & Plugins
Due to [1] the task 'Create data cinder-volumes group' is currently
failing when run on a restored image due to the idempotence failure.

In this patch we enable pinning the requirements file and plugins
repo to a different OSA branch, and set it to stable/rocky which is
currently pinned to Ansible 2.5.10 and does not exhibit this issue.

[1] https://github.com/ansible/ansible/issues/47301

Change-Id: Id0f98f3ea0f53173dae800d45f1a4554c8a0c38a