Add missing prometheus port for ceph auto discover
When ceph prometheus metrics are auto discovered the metricbeat config
should point to the ceph mgr prometheus port. Adding missing
brackets around metricset so the default is treated as an array.
Dropping ceph dir detection for prometheus auto discover and relying on
is the port availability and inventory group only.

Change-Id: Iaba0fdece00414e17bc172f39e624374a9d273e8