MNAIO: Recovery the galera cluster with no gvwstate.dat file
On occasion, a set of saved images may not have a gvwstate.dat file.
Given that the MNAIO tooling is for test purposes only, we can take
the risk of just making one up in order to ensure that the cluster
starts on boot. DO NOT DO THIS IN PRODUCTION!

Change-Id: I1b055ef0d02d77afe03d1a90baa2e6890e986c32