Add a default retention policy for skydive indexes
Given the ops tools now have a skydive deployment capability there needs
curator needs to be able to detect the addition of skydive indexes and
build a curator policy accordingly.

This change adds the new retention policy to the overlay inventory
providing a sane default for most environments.

The retention action files have been updated to remove the "-" as an
index separator. This was done because not all indexes use a dash as a
divider.

Change-Id: I5b61720f27da00e0c3b92341355b09ea6c01caba