Update the galera_server role and change the mysqlcheck port
This change updates the galera deployment options for kolide which will
allow it to co-exist on the same host as an elasticsearch environment.

Because the galera cluster for kolide is on the backend, and not user
facing the installation and setup of a VIP for the backend galera
cluster is not needed.

Change-Id: Ia5f581ebb028bf35c00dc7606b0eb1a12ad54670
Signed-off-by: Kevin Carter <kevin.carter@rackspace.com>