Omit dahsboard on elk setup by default
With the introduction of the "infrustructure" panel and "canvas" becoming
stable, there's not a lot of reason to import the  general beat dashboards.
The default dashboard are almost always in a state of disrepare and take a
long time to import on high traffic clusters.

This channge removes the default dashboard from the beat setup role by
default. If a deployer wishes to renable the default dashboards, or add any
other beat flags, the variable `elastic_setup_flags` can be used to extend
the setup.

Change-Id: If44845f53e4d0cb1e91ec804060316fb852b4bfa
Signed-off-by: Kevin Carter <kevin@cloudnull.com>