Ensure the default version of Java is set
When installing ES and Logstash the system version of java needs to
match the expected version of java ES and Logstash will use. This
change, uses the `update-alternatives` command to set the java version
to the expected value when more than one java exists on a system.
The nessisity for this change came from OS level upgrades within
environments running OLD versions of ES. Upon upgrading the base OS
our playbooks could not complete an upgrade of ES which was due to the
java expectations. Once the alternantives were set accordingly the
upgrade completed without issues.

Change-Id: I9025967f723ee17940e11789f503e342cdad6f2a
Signed-off-by: cloudnull <kevin@cloudnull.com>