Fix plugins loading in embedded ansible
The connection plugin is being loaded, but the strategy plugin is not.
This breaks lookups for containers, so the physical host SSH connection
IP cannot be detected properly and attempts to hit the inventory_hostname
for physical_host SSH instead of the calculated ansible_host passed
from the strategy plugin.

Change-Id: Iff60a57a50d9257a09ce7ec2e6aa27e0e01664cd