doc: drop release notes + fix sphinxmark
The ops repo isn't really software.   It's a collection of playbooks
that we share between operators, therefore, release notes are a bit
confusing because there are so many playbooks inside of it.

They're also broken because of how we have a few links all over the
place.  There has been no release notes published either.

sphinxmark is no longer compatible with the latest release of Sphinx
which is causing all of our documentation jobs to fail.  This patch
removes it as our current usage of openstacktheme for documentation
already provides watermarks for current branch and notices for which
branch the documentation covers.

Change-Id: I521647c999e7dc17ba27f02bc5bccd422d97ab9a
Closes-Bug: #1824664