Updates to Grafana role
The inclusion of the grafana submodule contradicts the deployment
process in the README, where it is explicitly cloned to
/etc/ansible/roles. The empty folder for the submodule takes
higher precendence than explicit clone, meaning that the grafana
role is found, but contains no tasks. The submodule has been
deleted to stop this.

There are also failures when running behind a proxy in the
traefik-common role, These are now avoided by running the role
with deployment_environment_variables set.

Change-Id: Icc34087e17906c763286b98591bade73fa7c022e