Expose Elasticsearch data path configuration
Currently the location where elasticsearch stores its data is hardcoded
in the elasticsearch.yml configuration template to /var/lib/elasticsearch.
This commit allows the option to be configured, and exposes the value as a
default variable. This not only allows the data directory to be changed,
but also for elasticsearch to be configured to use multiple data directories
through a comma separated list. The default behaviour is unchanged.

Change-Id: I21dbcaeab997362f94e245434cc5aa35136aeee9