Explicitly use Elasticsearch data node for ILM index update
When installing/configuring beats, it may be the case that elastic-logstash
nodes are not running Elasticsearch directly, and so can't update the ILM
config accordingly. This commit explicitly delegates the ILM config update
to an elasticsearch data node, in order that this will always be possible.
As Elasticsearch and logstash are by default installed together, the default
behaviour is unchanged.

Change-Id: I6833585b9a02eddc3c9a1179e487cdeaefa30ac2