Update ILM policy if non-existent
Currently, ILM policies are not updated when running the elastic_ilm role if
the named ILM policy does not exist, and the beat itself is already installed.
This can occur if one changes the name of the ILM policy post-install using the
ilm_policy_name variable and wishes to update existing beats to reflect this.
This commit adds another task which explicitly pushes the ILM policy in this case.

Change-Id: I6b97af316843b9bc1feaa762d9074d824e2d02a8