Increase logstash output connection timeout for beats
When using Metricbeat in an environment where only the system module
is enabled, the configuration reports system metrics to logstash every
60s. However, the current connection timout is shorter than this,
leading to the connection resetting and re-opening repeatedly due to
i/o timeouts. This commit increases the timeout to 90s, such that the
connection stays open and errors aren't reported to the metricbeat
logs.

Change-Id: Ib30f983df71c0b0381bfa23d9d2c7f1220184a4b