Update journalbeat name to use hostname
Currently, journalbeat sets the beat name under which it appears in Kibana
to 'journalbeat'. This is unlike the other beats, which don't set the value,
and therefore default to using the hostname, whilst containing the beat.type
field denoting the beat type. This commit changes the journalbeat
configuration to match the other beats, in order to standardise the output
data in Elasticsearch and Kibana.

Change-Id: Ib43c7e89e6b037f4a6c260f3375f491b65750bd4