Allow user override of beat install locations
The installation locations of elastic beats are currently hardcoded
using openstack-ansible groups and cannot readily be extended to include
non-openstack hosts. This is particularly notable with auditbeat,
packetbeat and filebeat, which directly use the OSA 'hosts' group, and
metricbeat, which uses 'all'.

This commit allows these choices to be overrided by defining groups
named filebeat, metricbeat etc, with naming conventions as with the elk
integration inventory. Using integration inventories, one can include
the OSA groups within the override to extend the deployment. The defaults
are unchanged, with the exception of Metricbeat, which is now by default
deployed to the OSA groups 'hosts' and 'all_containers' instead of 'all'.

Change-Id: Ie39fc10f5749c9636c890a0417fc532f77562144