Expose options for Kibana load balancing
In order to load balance between multiple Kibana instances, they must
share common encryption keys for cookies and Elasticsearch reports.
This commit allows these common keys to be optionally applied to Kibana.
If not specified, as is the default they aren't set, and Kibana defaults
to unique per-instance keys as before.

Change-Id: I873e9d22d3c6bf687305c6620097b9140df72f89