Exposed Kibana elastic endpoint
Kibana can currently only be deployed alongside an elasticsearch
coordinator node, so that it can query the cluster directly. However, in
certain situations it can be useful to keep kibana separate from
elasticsearch for e.g. resource reasons. This change allows the user to
deploy Kibana with a list of remote endpoints that it can use to query
the elastic cluster, such that no local install is required. The default
behaviour is unchanged, as the default endpoint remains the localhost.

Change-Id: Ic99b9aafd1d70b99d2088a4017e58079b5124ddc