Subdivision of elastic-logstash group
It is currently assumed that the elasticsearch data nodes and the
logstash nodes are co-located within the same container during
deployment of the ELK stack. This commit is intended to remove that
assumption by requiring specification of separate elastic and
logstash groups. This is particularly important where instances of
elasticsearch and logstash are separated for resource reasons, a
common setup for an elastic cluster.

The default setup for the elastic nodes has been simplified, with
all non-coordinating nodes being both data and ingest nodes unless
otherwise specified. This was previously defined algorithmically,
with the elastic node setup tightly coupled to the placement of the
logstash instances.

Change-Id: Id06eb78b52705aefea9cfe1247f53bac58badd52