Use service_facts to check haproxy existence
It is currently the case that at deploy time, the existence of a
/etc/haproxy directory is used to check whether or not haproxy is
running. This is used to check whether or not to enable the haproxy
module. This will cause frequent errors on metricbeat if haproxy is
installed but not running, as the endpoint will not be active. This
commit uses the service_facts modules to detect whether or not the
service is running at deploy time, and therefore whether or not to
enable the module.

Change-Id: If0e513e04d3dc567cedd16b440dc2771fddbbb5d