Use service_facts to check mdmonitor existence
Currently metricbeat will always enable checks for raid as part of the
system module metricset for physical hosts. This causes error logs and
generates noise for physical hosts without raid enabled. This change
checks that the mdmonitor.service is running before adding a raid check
to the system module metricset.

Change-Id: I2b96c30d2eec2c83787035a100a0fa29aedbf333