Expose Elasticsearch discovery nodes
It is currently the case that the discovery information for
elasticsearch nodes is generated using the elasticsearch node and
kibana node groups internally. This means that one cannot connect to
a remote cluster that is not managed by the same ansible inventory. This
commit exposes the configuration to allow this to occur, permitting
the elasticsearch discovery nodes to be optionally defined with further
information. The associated internal variables have been renamed from
zen_nodes as zen is used within Elasticsearch for discovery only
internally, and the associated elasticsearch config file setting has
been renamed to 'discovery.seed_hosts'.

Change-Id: I76532c185e991eddcb1111fc58e7120a52c0120c