Expose apt repo configuration
Currently one can not override the apt repository used for ELK stack
package installation due to variable precedence as the repo info is
defined in vars rather than defaults. To be able to do so is
particularly useful in the case that on wishes to specify alternative
mirrors to use. This commit optionally exposes the remote apt
configuration in the elastic_repoitory role defaults so that the user
can modify it in order to redirect to a mirror. The default
configuration is unchanged.

Change-Id: I6459bd8b1b55253a869b018d187c7ec29957a3c1