Allow non-url Elastic gpg key specification
It is currently the case that the elastic GPG key must be specified
as a remote URL within the elastic_repositories role. This causes
issues in the case that remote URLs specifying the GPG key are
inaccessible due to e.g. firewalls at deploy time.

In analogy to Change-ID I7ac1a5e3a05aa3d0b4fae86c4a325ef147a9a528,
this commit allows the GPG key to be provided not only via a remote
URL, but also in-data or through a file by allowing the full range
of apt_key input types (url, file, data etc) to be provided. The
default behaviour is changed to use the vendor key in the role
files.

Change-Id: Ic48db01029c4b94845ccacfba7440b13a59ab873