Added IPTables module support to Filebeat config
The IPtables module is currently absent in the filebeat config, meaning
that IPtables log data cannot be parsed into fields by filebeat. This
commit adds the appropriate configuration section to allow this to be
optionally enabled and configured with other iptables logfile locations.
The module is disabled by default, as filebeat will crash if the
appropriate log files are not found.

Change-Id: Ibd44234fd736d70e9e0bd5c4c6cab45f907c6d3c