Add Journalbeat support for OS LXC containers
Journalbeat currently only parses systemd journals
with machine UUID corresponding to the host, requiring
a separate journalbeat install to each OS container,
even when the container journals are present on the
host as subdirectories of /var/log/journal/.

This commit searches for these directories and adds
them to the Journalbeat journal path such that the
container journals are processed by the Journalbeat
instance on the physical host. This therefore removes
the need for Journalbeat to be deployed to each LXC
container.

Change-Id: Ifdfc8c52364d02fb2a0ca296b8c41c3d32bc0342