Correctly enable mysql filebeat module
There is currently an issue where the mysql module setup in the
filebeat config file will apply to the last enabled module due to
a misplaced comment. This will cause Filebeat to crash when both
IPTables and galera are present, as the error config section does
not exist. This commit uncomments the mysql module definition
in order for the configuration to apply correctly to both IPTables
and mysql.

Change-Id: Iecd41881b3adec4e4196c48334b6234c15e80f6c