Expose Journalbeat queue configuration
Currently the settings for the internal queue in journalbeat are left at their
default settings and unchangeable. This can result in a situation whereby logs
can remain in the queue for hours if there are not enough journal events
produced to trigger a flushing of the queue. One can observe this in kibana
through a large discrepancy between the output event.created and @timestamp
fields. This commit exposes the configuration options for the in-memory queue
which is used by default in journalbeat, such that the queue flushing can be
set with more appropriate timeouts etc. This commit leaves each value at its
current default.

Change-Id: I777c68a54226045966416d12d5bba56cb7eaeb68