Expose voting-only option for master-eligible elasticsearch nodes
Master-eligible Elasticsearch nodes can be prevented from winning
elections by enabling voting_only mode. This is particularly useful
where certain nodes need to be master-eligible for cluster
discovery and formation, but shouldn't be able to actually become
the master node for the entire cluster. This commit exposes the
option in the elasticsearch role to set this. The default behaviour
remains the same, with all master-eligible able to win elections.

Change-Id: Ieba965af57c86e3b3767444883114f0130d3f000