Alter default shard count in index templates
Currently, we use only a single primary shard for each beat index,
as configured by the index templates. This is not usually an optimal
setup, as all index searching is handled on a single node. A more
efficient configuration for a small number of different indices is
to balance, with around one primary shard per node. This
commit exposes the config option to set the number of primary shards,
which was previously hardcoded to 1. This now defaults to one shard
per data node in the cluster.

Change-Id: Icacf76c2270c98f8676d57952c00773646ca7b51