Allow minimal auditbeat deploy on LXC containers
Auditbeat currently cannot be deployed on containers, as the auditd
module requires auditd, which is not available on LXC containers.
However, the other modules - file_integrity, system do not have auditd
as a dependency. As a result, Auditbeat can be deployed more minimally
on containers if the auditd module is not used, providing filesystem
integrity checks, login, user, host and other data. The system socket
dataset also must be disabled, as this requires data not accessible from
the container.

This commit disables the auditd and system socket modules if the host
is containerised, allowing Auditbeat to be deployed to containers.

Change-Id: I96c92d2ce31d111ba5d7645b959ca287afbebd0b