Ensure that the beat setup role runs when elk_beat_setup=true
Currently it is possible for the elk_beat_setup flag not to trigger
the beat_setup role when no kibana nodes are present. This means that
there is no way to update the index templates in this situation, so
beats will just use a default index setup, and ILM policies will not
be applied. This can lead to inconsistent configurations where manual
deletion of default-templated indices is required. This can happen
during an ELK version upgrade, where new templates should be created
but aren't.

This commit ensures that the beat setup role is run whenever the
elk_beat_setup flag is set to true. As a result, the user can force
index templates to be updated even when no kibana nodes are present
to ensure consistency. It also documents the previously undocumented
elk_beat_setup var in both the Beat role defaults and the beat_setup
role itself

Change-Id: Ife0d0f12d6b300b84b63c3af14add1732747d44f