Add override for Auditbeat socket dataset
When deployed to a container host, the socket dataset of the Auditbeat
system module can become very CPU-intensive, as it captures all socket
events from all containers, data which largely overlaps with data
collected using Packetbeat. This commit adds an override flag to allow
this dataset to be selectively disabled in the case that this CPU usage
becomes too high.

Change-Id: I595c6b909a83e6f49b1042dff29b38953768ef67