Prevent duplicate initial_master_nodes
Elasticsearch fails to start on first run when duplicate entries in the
cluster.initial_master_nodes variable are present, as they are all required
during the initial bootstrapping. It is currently possible for duplicates
to be present in this when nodes are forced to be master-eligible using
elasticsearch_node_master: true
due to the master node setting algorithm within data-node-variables. This
commit ensures that all values within the Ansible master_nodes variable
are unique, whilst keeping the total count of master nodes the same,
preventing Elasticsearch from crashing during initial cluster discovery.

Change-Id: I3763b22ea47c9b5b0f534793ccc6a887f96dc649