Fix regex_replace mapping
With current mapping we result in elasticsearch_data_node_details being
172.29.236.126:9200:9200 instead of 172.29.236.126:9200 which
means that map is applied twice to the same item
Changing regexp helps to resolve it. It looks like related to what [1] have seen

Also we don't need to transform to list when doing another map, as
working with generator object is more resource efficient.

[1] https://gist.github.com/halberom/9bfe009c9a3df5d64c7bb505a7700430#gistcomment-2706901

Change-Id: I22b6735ad211e60b61d74cea34c7987d74bc8596