Respect defined version and source of mcapi driver
At the moment proxy and main mcapi versions are not related in
any way, which means that there can be a conflict between
installed versions of proxy with mcapi.

One of examples was triggered by [1], when proxy was issuing
a requests towards API which does not exist yet in the central
cluster. This resulted in HTTPError: 404 Client Error: Not Found for url
when proxy tried to fetch list of clusters.

[1] https://github.com/vexxhost/magnum-cluster-api/commit/8e375c414ab3e2027486b105b12765891850c8ac

Change-Id: Iba1f34e7b79e34f23b8169ecd20907832626d540